Folder Hierarchy redone, Main scene worked on.
the tutorial project we were following gave us all of the Asset folders'  data, so we rearranged it to not conflict with the folder system Unity uses. Also for peace of mind.
The build works, Main goes to level goes to level1 goes to level2. Weapons, enemies, camera, buying, upgrading, and placing all work.